【修复】修复servo执行器的int32强转以修复warning

+++ code/servo_executor.c
@@ -81,10 +81,11 @@ void servo_executor_update(void)
 
     // 4. 【核心】使用斜率限制，平滑地趋近目标值
     // 公式: new = last + limit(target - last, -dec, +acc)
-    int32 current_duty_lf = PWM_CH1_LAST + Float_Constrain(target_final_duty_lf - PWM_CH1_LAST, -dec_limit, acc_limit);
-    int32 current_duty_rf = PWM_CH2_LAST + Float_Constrain(target_final_duty_rf - PWM_CH2_LAST, -dec_limit, acc_limit);
-    int32 current_duty_rr = PWM_CH3_LAST + Float_Constrain(target_final_duty_rr - PWM_CH3_LAST, -dec_limit, acc_limit);
-    int32 current_duty_lr = PWM_CH4_LAST + Float_Constrain(target_final_duty_lr - PWM_CH4_LAST, -dec_limit, acc_limit);
+    int32 current_duty_lf = PWM_CH1_LAST + (int32)Float_Constrain(target_final_duty_lf - PWM_CH1_LAST, -dec_limit, acc_limit);
+    int32 current_duty_rf = PWM_CH2_LAST + (int32)Float_Constrain(target_final_duty_rf - PWM_CH2_LAST, -dec_limit, acc_limit);
+    int32 current_duty_rr = PWM_CH3_LAST + (int32)Float_Constrain(target_final_duty_rr - PWM_CH3_LAST, -dec_limit, acc_limit);
+    int32 current_duty_lr = PWM_CH4_LAST + (int32)Float_Constrain(target_final_duty_lr - PWM_CH4_LAST, -dec_limit, acc_limit);
+
 
     // 5. 更新内部状态，为下一次计算做准备
     PWM_CH1_LAST = current_duty_lf;
